Execption handling

+++ src/main/java/com/devsuperior/dacommerce/services/ProductService.java
@@ -5,14 +5,20 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dacommerce.DTO.ProductDTO;
 import com.devsuperior.dacommerce.entities.Product;
 import com.devsuperior.dacommerce.repositories.ProductRepository;
+import com.devsuperior.dacommerce.services.exceptions.DataBaseException;
+import com.devsuperior.dacommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -22,7 +28,7 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		ProductDTO dto = new ProductDTO(repository.findById(id).get());
+		ProductDTO dto = new ProductDTO(repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado")));
 		return dto;
 	}
 	
@@ -43,17 +49,28 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto){ 
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto,entity);
-		entity = repository.save(entity);
-		
-		return new ProductDTO(entity);
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto,entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		}
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id){ 
-		repository.deleteById(id);;
-	
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);;
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DataBaseException("Falha de integridade referencial");
+		}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
